added tests for the calculation logic and minor clean up

+++ app/src/main/java/club/premiumit/consentlab/domain/model/Calculation.kt
@@ -2,5 +2,5 @@ package club.premiumit.consentlab.domain.model
 
 data class Calculation(
     val dataTypes: List<DataType>,
-    val totalCost: Double
+    val totalCost: Int
 )
